+++ dopus4-os4/trunk/library-os4/main/CheckExist.c
@@ -68,7 +68,7 @@ int _DOpus_CheckExist(struct DOpusIFace *Self, char *name, int *size)
 	{
 		if(size)
 		{
-			*size = data->FileSize;
+			*size = (int)data->FileSize;
 		}
 		if(EXD_IS_FILE(data))
 		{
